+++ test/test_jet.jl
@@ -1,26 +1,29 @@
 @testitem "JET checks" tags=[:jet] begin
-using JET
-using ArrayInterface
-using Static
-using Graphs
-using StridedViews
-using LinearAlgebra
-using AbstractAlgebra, Hecke
+    using JET
+    using Test
+    using ArrayInterface
+    using Static
+    using Graphs
+    using StridedViews
+    using LinearAlgebra
+    using Nemo
+    using AbstractAlgebra
+    using Hecke
 
+    rep = report_package("QuantumClifford";
+        ignored_modules=(
+            AnyFrameModule(Graphs.LinAlg),
+            AnyFrameModule(Graphs.SimpleGraphs),
+            AnyFrameModule(ArrayInterface),
+            AnyFrameModule(Static),
+            AnyFrameModule(StridedViews),
+            AnyFrameModule(LinearAlgebra),
+            AnyFrameModule(Nemo),
+            AnyFrameModule(AbstractAlgebra),
+            AnyFrameModule(Hecke),
+    ))
 
-rep = report_package("QuantumClifford";
-    ignored_modules=(
-        AnyFrameModule(Graphs.LinAlg),
-        AnyFrameModule(Graphs.SimpleGraphs),
-        AnyFrameModule(ArrayInterface),
-        AnyFrameModule(Static),
-        AnyFrameModule(StridedViews),
-        AnyFrameModule(LinearAlgebra),
-        AnyFrameModule(Hecke),
-        AnyFrameModule(AbstractAlgebra),
-    )
-)
-@show rep
-@test_broken length(JET.get_reports(rep)) == 0
-@test length(JET.get_reports(rep)) <= 23
+    @show rep
+    @test_broken length(JET.get_reports(rep)) == 0
+    @test length(JET.get_reports(rep)) <= 11
 end