+++ test/test_symcliff.jl
@@ -66,7 +66,7 @@
     end
 
     @testset "SingleQubitOperator inv methods" begin
-        for gate_type in [sHadamard, sX, sY, sZ, sId1 , sPhase, sInvPhase]
+        for gate_type in filter(gate_type -> gate_type != SingleQubitOperator, subtypes(AbstractSingleQubitOperator))
             n = rand(1:10)
             @test CliffordOperator(inv(SingleQubitOperator(gate_type(n))), n) == inv(CliffordOperator(gate_type(n), n))
             @test CliffordOperator(inv(gate_type(n)), n) == inv(CliffordOperator(gate_type(n), n))
@@ -76,6 +76,17 @@
             @test CliffordOperator(inv(random_op), i) == inv(CliffordOperator(random_op, i))
             @test CliffordOperator(inv(SingleQubitOperator(random_op)), i) == inv(CliffordOperator(random_op, i))
         end
+
+    @testset "Consistency checks with Stim" begin
+       # see https://github.com/quantumlib/Stim/blob/main/doc/gates.md
+       @test CliffordOperator(sCXYZ)       == C"Y X"
+       @test CliffordOperator(sCZYX)       == C"Z Y"
+       @test CliffordOperator(sSQRTX)      == C"X -Y"
+       @test CliffordOperator(sSQRTY)      == C"-Z X"
+       @test CliffordOperator(sInvSQRTX)   == C"X Y"
+       @test CliffordOperator(sInvSQRTY)   == C"Z -X"
+       @test CliffordOperator(sHadamardXY) == C"Y -Z"
+       @test CliffordOperator(sHadamardYZ) == C"-X Y"
     end
 
     @testset "TwoQubitOperator inv methods" begin