+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ## dev
 
+- Implementing many more named single-qubit gates following naming convention similar to the stim package in python.
 - **(fix)** Bug fix to the `parity_checks(ReedMuller(r, m))` of classical Reed-Muller code (it was returning generator matrix).
 - `RecursiveReedMuller` code implementation as an alternative implementation of `ReedMuller`.